test(annotations): verify visual choices at actual icon size
Move renderer-backed preview geometry into a testable shared helper and exercise every fill, pattern, width, shape, route and arrowhead preview in the existing harness.
